+++ Tests/SwiftLintFrameworkTests/CommandTests.swift
@@ -1,6 +1,7 @@
 // swiftlint:disable file_length
 import Foundation
 import SourceKittenFramework
+@testable import SwiftLintBuiltInRules
 @testable import SwiftLintCore
 import XCTest
 
@@ -453,5 +454,30 @@ class CommandTests: SwiftLintTestCase {
             )),
             []
         )
+    }
+
+    func testSuperfluousDisableCommandsEnabledForAnalyzer() {
+        let configuration = Configuration(
+            rulesMode: .default(disabled: [], optIn: [UnusedDeclarationRule.description.identifier])
+        )
+        let violations = violations(
+            Example("""
+            public class Foo {
+                // swiftlint:disable:next unused_declaration
+                func foo() -> Int {
+                    1
+                }
+                // swiftlint:disable:next unused_declaration
+                func bar() {
+                   foo()
+                }
+            }
+            """),
+            config: configuration,
+            requiresFileOnDisk: true
+        )
+        XCTAssertEqual(violations.count, 1)
+        XCTAssertEqual(violations.first?.ruleIdentifier, "superfluous_disable_command")
+        XCTAssertEqual(violations.first?.location.line, 3)
     }
 }